more organization and refactoring (i'm overthinking oh no)
