+++ config.json
@@ -1,0 +1,4 @@
+{
+    "baseUrl": "https://discord.com/api/v",
+    "apiVersion": "10"
+}
