+++ src/templates/discord/ApplicationCommand.ts
