+++ src/templates/discord/ChannelResources.ts
