+++ src/templates/discord/Embed.ts
