+++ src/templates/discord/EmojiResources.ts
