+++ src/templates/discord/Enums.ts
