+++ src/templates/discord/InteractionResponse.ts
