+++ src/templates/discord/UserResource.ts
